pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Django project with microservices"
authors = [
    {name = "rvishakhs",email = "rvishakhs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
